Multi attesters (#367)
* feat(core): add sync attester

* chore: lint

* feat(core): init with multiple attesters

* feat(core): documentation for multiple attesters in synchronizer

* fix(core): fix input type of userState

* feat(core): use one `_attesterId` array

* feat(core): sync states independently of attesterID

* chore(core): revert calcCurrentEpoch and getEpochKeys functions

* chore(core): remove private stateTree

* chore(core): add helper functions, add error descriptions

* feat(core): should calculate epoch while syncing all attesters

* fix(core): update @semphoare-procotol/identity

* chore(core): use filter to query events once

* chore: revert Synchronizer State with attester IDs

* chore(core): add error messages

* fix(core): epoch key indexes, to decimal attester ID

* fix(test): remove genSealedEpochProof, ust nullifier

* chore: lint

* chore(core): add synchronizer stop/restart test

* fix(core): fix synchronizer stop/restart test

* chore(core): check number of items in db

* fix(core): idempotent handlers

---------

+++ packages/core/test/userState.test.ts
@@ -64,7 +64,7 @@ describe('User state', function () {
             .connect(attester)
             .userSignUp(publicSignals, proof)
             .then((t) => t.wait())
-        const epk = userState.getEpochKeys(epoch, 1)
+        const epk = userState.getEpochKeys(epoch, 1) as bigint
         const fieldIndex = 1
         const v0 = F - BigInt(1)
         const v1 = BigInt(12409124)
